sbir-etl/pyproject.toml
[tool.poetry]
name = "sbir-loader"
version = "0.1.0"
description = "Fetch SBIR data and load it into Postgres"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "sbir_loader" }]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
httpx = "^0.27.0"
pydantic = "^2.7.0"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pgvector = "^0.4.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
alembic = "^1.16.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"